+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Handle NPE inside of `JobScheduler`
 * Handle 'Package manager has died' crash
+* Save the highest job ID in a pref file so that it isn't lost when no job is in the database anymore (makes debugging easier)
 
 ## 1.1.11 (2017-06-05)